pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-qpay"
version = "1.0.0"
description = "QPay V2 payment integration for Django"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = ["qpay-py>=1.0", "django>=4.2"]

[project.urls]
Repository = "https://github.com/qpay-sdk/django-qpay"
